🎨 Format invalide component
