+++ packages/applications/ssr/src/components/pages/réseau/raccordement/détails/DétailsRaccordement.page.tsx
@@ -94,15 +94,3 @@ export const DétailsRaccordementPage: FC<DétailsRaccordementPageProps> = ({
     </Button>
   </PageTemplate>
 );
-
-export const FormatFichierInvalide: FC = () => (
-  <div className="flex items-center gap-1">
-    <Icon
-      id="fr-icon-alert-fill"
-      size="sm"
-      className=" text-warning-425-base"
-      title="format du fichier invalide"
-    />
-    <p className="text-xs">Le format du fichier est invalide</p>
-  </div>
-);
